refactor package per review
