+++ bundles-maven-plugin/src/main/java/org/apache/metron/maven/plugins/bundles/BundleMojo.java
@@ -14,7 +14,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.metron;
+
+package org.apache.metron.maven.plugins.bundles;
 
 import java.io.File;
 import java.io.IOException;
